pyproject.toml
[project]
name = "probetag-werki"
version = "0.1.0"
description = "Reference implementation of a Keycloak MCP server for student internship trial"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Vincenzo Pace", email = "pace@amiconsult.de" }
]
# Runtime dependencies - required to run the application
dependencies = [
    "fastmcp>=2.13.0.2",
    "pydantic>=2.10.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

# Optional dependencies for development
# Install with: uv pip install -e ".[dev]"
[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.4",
    "responses>=0.25.4",  # Mock HTTP requests

    # Type checking
    "mypy>=1.14.0",
    "types-requests>=2.32.0.0",

    # Code formatting and linting
    "ruff>=0.8.5",
]

# Tool configurations
[tool.pytest.ini_options]
# Configuration for pytest
testpaths = ["test"]
python_files = "test_*.py"
addopts = ["--verbose"]

[tool.mypy]
# Type checking configuration
python_version = "3.13"
ignore_missing_imports = true  # Simplify for learning - ignore missing stubs

[tool.ruff]
# Ruff is a fast Python linter and formatter
line-length = 100
target-version = "py313"

[tool.ruff.lint]
# Enable basic linting rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes (unused imports, undefined names)
]
